Power BI Loan Project
A loan is a form of debt incurred by an individual or other entity. The lender—usually a corporation, financial institution, or government—advances a sum of money to the borrower. Power BI Loan Project showcases the total number of customers, money borrowed, and the number of collateral obtained by a bank.